refactor(cli): reorganize spec/impl pipeline around its data entities
Apply the "Life of a File" rule to the CLI's spec/impl pipeline:

- Extract bundler/esbuild logic into Bundler.ts (previously sprawled
  across utils.ts and implValidation.ts), exposing a single `bundle`
  entry point plus `directlyImports` for metafile queries.
- Rename modulePaths.ts to LeafModule.ts and fold validateSpec /
  validateImpl into it, since they operate on a LeafModule value.
- Inline validateSchema into confect/codegen.ts (single caller).
- Delete implValidation.ts.
- Replace specAssembly.ts with SpecAssemblyNode.ts: hide the mutable
  SpecTreeNode intermediary, expose only assemblyNodesFromLeaves and
  partitionByRuntime, and build the assembly tree purely.
- Reorganize tests around the new modules (LeafModule.test.ts,
  SpecAssemblyNode.test.ts, codegen.test.ts).

No behavior changes.
